// src/api-generated/model/createRoleDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaS API
 * The core API for my multi-tenant SaaS
 * OpenAPI spec version: 1.0
 */
import type { CreateRoleDtoNameTranslations } from './createRoleDtoNameTranslations';

export interface CreateRoleDto {
  /**
   * The Turkish name of the role (primary)
   * @minLength 3
   * @maxLength 255
   */
  name: string;
  /** Translations of the role name (optional) */
  name_translations: CreateRoleDtoNameTranslations;
  /**
   * Array of permission IDs to assign to the role
   * @minItems 1
   */
  permissionIds: string[];
}