// src/api-generated/model/updateRoleDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaS API
 * The core API for my multi-tenant SaaS
 * OpenAPI spec version: 1.0
 */
import type { UpdateRoleDtoNameTranslations } from './updateRoleDtoNameTranslations';

export interface UpdateRoleDto {
  /**
   * The updated Turkish name of the role
   * @minLength 3
   * @maxLength 255
   */
  name?: string;
  /** Updated translations of the role name */
  name_translations: UpdateRoleDtoNameTranslations;
  /**
   * Updated array of permission IDs to assign to the role
   * @minItems 1
   */
  permissionIds?: string[];
  /** Permission IDs to remove from the role incrementally */
  permissionsToRemove?: string[];
  /** Permission IDs to add to the role incrementally */
  permissionsToAdd?: string[];
}